feat: implemented a dropdown to test the library
also added some notes about problems with the custom ID system. Probably better to just have the user input the ID themselves, rather than automatically calculating it. The persistent data system will also be more stable this way, as it's the users responsibility to avoid duplicate ids, not ours. If we want auto id handling, that should be done at the Paper level, not the component library level